More fixing of warnings after noticing clang was skipping some tests.

+++ examples/temp_sensor/src/UsartPutChar.c
@@ -18,6 +18,6 @@ void Usart_PutChar(char data)
 #ifdef SIMULATE
   printf("%c", data);
 #else
-  AT91C_BASE_US0->US_THR = data;
+  AT91C_BASE_US0->US_THR = (uint8)data;
 #endif
 }
